+++ controllers/recipesController.js
@@ -1,15 +1,21 @@
+const { validationResult } = require('express-validator');
 const Recipe = require('../models/recipe');
 
-exports.getAllRecipes = async (req, res) => {
+exports.getAllRecipes = async (req, res, next) => {
   try {
     const recipes = await Recipe.find();
     res.status(200).json(recipes);
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
 
-exports.getRecipeById = async (req, res) => {
+exports.getRecipeById = async (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
     const recipe = await Recipe.findById(req.params.id);
     if (!recipe) {
@@ -17,21 +23,31 @@ exports.getRecipeById = async (req, res) => {
     }
     res.status(200).json(recipe);
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
 
-exports.createRecipe = async (req, res) => {
+exports.createRecipe = async (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
     const newRecipe = new Recipe(req.body);
     await newRecipe.save();
     res.status(201).json(newRecipe);
   } catch (error) {
-    res.status(400).send(error);
+    next(error);
   }
 };
 
-exports.updateRecipe = async (req, res) => {
+exports.updateRecipe = async (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
     const recipe = await Recipe.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
     if (!recipe) {
@@ -39,11 +55,16 @@ exports.updateRecipe = async (req, res) => {
     }
     res.status(200).json(recipe);
   } catch (error) {
-    res.status(400).send(error);
+    next(error);
   }
 };
 
-exports.deleteRecipe = async (req, res) => {
+exports.deleteRecipe = async (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
     const recipe = await Recipe.findByIdAndDelete(req.params.id);
     if (!recipe) {
@@ -51,6 +72,6 @@ exports.deleteRecipe = async (req, res) => {
     }
     res.status(200).json({ message: 'Recipe deleted' });
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
